Add new bundled git tests and update existing ones
